Post release fixes
