+++ apps/backend-functions/src/main.ts
@@ -144,11 +144,13 @@ export const twilioWebhook = functions().https.onRequest(_twilioWebhook);
 /**
  * Creates notifications when an event is about to start
  */
-export const finishedScreenings = functions().pubsub.schedule('*/30 * * * *') // every 30 minutes
-  .onRun(skipInMaintenance(createNotificationsForFinishedScreenings));
  
 export const scheduledNotifications = functions().pubsub.schedule('*/30 * * * *') // every 30 minutes
-  .onRun(skipInMaintenance(createNotificationsForEventsToStart));
+  .onRun(skipInMaintenance(() => {
+    createNotificationsForEventsToStart();
+    createNotificationsForFinishedScreenings();
+  }));
+
 
 //--------------------------------
 //       Movies Management      //
