+++ apps/crm/crm/src/app/crm/components/guest-table/guest-table.component.ts
@@ -1,6 +1,5 @@
 import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
-import { getGuest } from '@blockframes/invitation/pipes/guest.pipe';
-import { Invitation, InvitationDetailed } from '@blockframes/model';
+import { getGuest, Invitation, InvitationDetailed } from '@blockframes/model';
 
 @Component({
   selector: 'invitation-guest-table',
