+++ apps/crm/crm/src/app/crm/pages/invitations/invitations.component.ts
@@ -4,8 +4,7 @@ import { OrganizationService } from '@blockframes/organization/service';
 import { EventService } from '@blockframes/event/service';
 import { downloadCsvFromJson } from '@blockframes/utils/helpers';
 import { getHost } from '@blockframes/invitation/pipes/host.pipe';
-import { getGuest } from '@blockframes/invitation/pipes/guest.pipe';
-import { Movie, Organization, orgName, Event, isScreening, InvitationDetailed } from '@blockframes/model';
+import { Movie, Organization, orgName, Event, isScreening, InvitationDetailed, getGuest } from '@blockframes/model';
 import { MovieService } from '@blockframes/movie/service';
 import { where } from 'firebase/firestore';
